Translate initializers by sequencing them all
I realised we cannot translate Ivy initializers separately. At least
in Ivy 1.8 (and I think 1.7 as well), initializers are now stateful
and semanticaly they execute one after the other, such that the changes
made by former initializers are visible to the latter.

To account for this, we create an 'artificial' action that Sequences
all the `after init` block in the order they show up in
`mod.initializers`, and translate this action as a single initializer.

Whereas actions have parameters, for initializers we explicitly
add an outer existential quantifier. Moreover, we use `make_vc()`
rather than `update()` to generate a one-state formula, as opposed
to the two-state formulas normally used for actions.